requirements.txt
absl-py==0.12.0
altair==4.1.0
apache-beam==2.38.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
blinker==1.4
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
clang==5.0
click==7.1.2
cloudpickle==2.0.0
crcmod==1.7
Cython==0.29.28
dataclasses==0.8
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.1.1
docker==4.4.4
docopt==0.6.2
entrypoints==0.4
fastavro==1.4.7
fasteners==0.17.3
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.18
google-api-core==1.31.5
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.12.1
google-cloud-bigquery==2.34.3
google-cloud-bigquery-storage==2.13.1
google-cloud-bigtable==1.7.1
google-cloud-core==1.7.2
google-cloud-datastore==1.15.4
google-cloud-dlp==3.6.2
google-cloud-language==1.3.1
google-cloud-pubsub==2.12.0
google-cloud-pubsublite==1.4.1
google-cloud-recommendations-ai==0.2.0
google-cloud-resource-manager==1.4.1
google-cloud-spanner==1.19.2
google-cloud-storage==2.0.0
google-cloud-videointelligence==1.16.2
google-cloud-vision==1.0.1
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
grpc-google-iam-v1==0.12.4
grpcio==1.44.0
grpcio-gcp==0.2.2
grpcio-status==1.44.0
h5py==3.1.0
hdfs==2.7.0
httplib2==0.19.1
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.17.2
Jinja2==3.0.3
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
keras-tuner==1.1.2
kt-legacy==1.0.4
kubernetes==12.0.1
Markdown==3.3.6
MarkupSafe==2.0.1
mistune==0.8.4
ml-metadata==1.3.0
ml-pipelines-sdk==1.3.4
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.5
notebook==6.4.10
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.2.0
opt-einsum==3.3.0
orjson==3.6.1
overrides==6.1.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.5.0
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
portpicker==1.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
proto-plus==1.20.3
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.6.2
pydot==1.4.2
Pygments==2.11.2
pymongo==3.12.3
Pympler==1.0.1
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
semver==2.13.0
Send2Trash==1.8.0
six==1.15.0
sklearn==0.0
smmap==5.0.0
streamlit==1.8.1
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.6.2
tensorflow-data-validation==1.3.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-model-analysis==0.34.1
tensorflow-serving-api==2.6.2
tensorflow-transform==1.3.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.6.0
tfx==1.3.4
tfx-bsl==1.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
typing-utils==0.1.0
tzdata==2022.1
tzlocal==4.2
uritemplate==3.0.1
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wrapt==1.12.1
zipp==3.6.0
